backend/requirements.txt
black==22.12.0
Django==3.2
django-filter==22.1
djangorestframework==3.12.0
djoser==2.1.0
flake8==5.0.4
gunicorn==20.1.0
Pillow==9.3.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
requests==2.28.1
virtualenv==20.17.1
